+++ archicad-addon/Examples/test.py
@@ -5,5 +5,5 @@ conn = ACConnection.connect ()
 acc = conn.commands
 act = conn.types
 
-response = acc.ExecuteAddOnCommand (act.AddOnCommandId ('TapirCommand', 'GetArchicadLocation'))
+response = acc.ExecuteAddOnCommand (act.AddOnCommandId ('TapirCommand', 'GetProjectInfo'))
 print (response)
