+++ archicad-addon/Sources/AddOn/AddOnMain.cpp
@@ -2,6 +2,8 @@
 #include "ACAPinc.h"
 
 #include "ResourceIds.hpp"
+#include "ApplicationCommands.hpp"
+#include "ProjectCommands.hpp"
 #include "Commands.hpp"
 
 API_AddonType __ACDLL_CALL CheckEnvironment (API_EnvirParams* envir)
@@ -23,7 +25,7 @@ GSErrCode __ACENV_CALL Initialize (void)
     err |= ACAPI_Install_AddOnCommandHandler (GS::NewOwned<TeamworkReceiveCommand> ());
     err |= ACAPI_Install_AddOnCommandHandler (GS::NewOwned<GetProjectInfoCommand> ());
     err |= ACAPI_Install_AddOnCommandHandler (GS::NewOwned<GetArchicadLocationCommand> ());
-    err |= ACAPI_Install_AddOnCommandHandler (GS::NewOwned<QuitCommand> ());
+    err |= ACAPI_Install_AddOnCommandHandler (GS::NewOwned<QuitArchicadCommand> ());
     err |= ACAPI_Install_AddOnCommandHandler (GS::NewOwned<ReloadLibrariesCommand> ());
     err |= ACAPI_Install_AddOnCommandHandler (GS::NewOwned<MoveElementsCommand> ());
     err |= ACAPI_Install_AddOnCommandHandler (GS::NewOwned<CreateColumnsCommand> ());
