+++ archicad-addon/Sources/AddOn/ApplicationCommands.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "CommandBase.hpp"
+
+class GetArchicadLocationCommand : public CommandBase
+{
+public:
+    virtual GS::String GetName () const override;
+    virtual GS::Optional<GS::UniString> GetResponseSchema () const override;
+    virtual GS::ObjectState Execute (const GS::ObjectState& parameters, GS::ProcessControl& processControl) const override;
+};
+
+class QuitArchicadCommand : public CommandBase
+{
+public:
+    virtual GS::String GetName () const override;
+    virtual GS::ObjectState Execute (const GS::ObjectState& parameters, GS::ProcessControl& processControl) const override;
+};
