+++ archicad-addon/Sources/AddOn/Commands.hpp
@@ -17,29 +17,6 @@ public:
     virtual GS::ObjectState Execute (const GS::ObjectState& parameters, GS::ProcessControl& processControl) const override;
 };
 
-class GetProjectInfoCommand : public CommandBase
-{
-public:
-    virtual GS::String GetName () const override;
-    virtual GS::Optional<GS::UniString> GetResponseSchema () const override;
-    virtual GS::ObjectState Execute (const GS::ObjectState& parameters, GS::ProcessControl& processControl) const override;
-};
-
-class GetArchicadLocationCommand : public CommandBase
-{
-public:
-    virtual GS::String GetName () const override;
-    virtual GS::Optional<GS::UniString> GetResponseSchema () const override;
-    virtual GS::ObjectState Execute (const GS::ObjectState& parameters, GS::ProcessControl& processControl) const override;
-};
-
-class QuitCommand : public CommandBase
-{
-public:
-    virtual GS::String GetName () const override;
-    virtual GS::ObjectState Execute (const GS::ObjectState& parameters, GS::ProcessControl& processControl) const override;
-};
-
 class ReloadLibrariesCommand : public CommandBase
 {
 public:
