+++ archicad-addon/Sources/AddOn/ProjectCommands.cpp
@@ -1,0 +1,71 @@
+#include "ProjectCommands.hpp"
+#include "ObjectState.hpp"
+
+GS::String GetProjectInfoCommand::GetName () const
+{
+    return "GetProjectInfo";
+}
+
+GS::Optional<GS::UniString> GetProjectInfoCommand::GetResponseSchema () const
+{
+    return R"({
+        "type": "object",
+        "properties": {
+            "isUntitled": {
+                "type": "boolean",
+                "description": "True, if the project is not saved yet."
+            },
+            "isTeamwork": {
+                "type": "boolean",
+                "description": "True, if the project is a Teamwork (BIMcloud) project."
+            },
+            "projectLocation": {
+                "type": "string",
+                "description": "The location of the project in the filesystem or a BIMcloud project reference.",
+                "minLength": 1
+            },
+            "projectPath": {
+                "type": "string",
+                "description": "The path of the project. A filesystem path or a BIMcloud server relative path.",
+                "minLength": 1
+            },
+            "projectName": {
+                "type": "string",
+                "description": "The name of the project.",
+                "minLength": 1
+            }
+        },
+        "additionalProperties": false,
+        "required": [
+            "isUntitled",
+            "isTeamwork"
+        ]
+    })";
+}
+
+GS::ObjectState GetProjectInfoCommand::Execute (const GS::ObjectState& /*parameters*/, GS::ProcessControl& /*processControl*/) const
+{
+    API_ProjectInfo projectInfo = {};
+    GSErrCode err = ACAPI_Environment (APIEnv_ProjectID, &projectInfo);
+
+    if (err != NoError) {
+        return CreateErrorResponse (APIERR_NOPLAN, "Failed to retrieve project information. Check the opened project!");
+    }
+
+    GS::ObjectState response;
+    response.Add ("isUntitled", projectInfo.untitled);
+    response.Add ("isTeamwork", projectInfo.teamwork);
+    if (!projectInfo.untitled) {
+        if (projectInfo.location) {
+            response.Add ("projectLocation", projectInfo.location->ToDisplayText ());
+        }
+        if (projectInfo.projectPath) {
+            response.Add ("projectPath", *projectInfo.projectPath);
+        }
+        if (projectInfo.projectName) {
+            response.Add ("projectName", *projectInfo.projectName);
+        }
+    }
+
+    return response;
+}
