+++ archicad-addon/Sources/AddOn/ProjectCommands.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "CommandBase.hpp"
+
+class GetProjectInfoCommand : public CommandBase
+{
+public:
+    virtual GS::String GetName () const override;
+    virtual GS::Optional<GS::UniString> GetResponseSchema () const override;
+    virtual GS::ObjectState Execute (const GS::ObjectState& parameters, GS::ProcessControl& processControl) const override;
+};
